オプション受け取りの処理を分割した
