+++ 08.ls_object/file_info.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 class FileInfo
   def initialize(file_name)
     @stat = File.stat(file_name)
